fix: deadlock when one thread race around the loop
The bug is caused by the thread who broadcasts other threads racing in the next round to reach the barrier while loop before the last thread exits the while loop, which will cause the last thread be stuck in the cond_wait.

I fixed this by adding a local variable to record the round number when enter the loop, if it is changed, that means another round should have begun, and the barrier condition has already met. So we can just skip the while loop if the round changed.